client -> execution refactor: added new execution module, new Execution abstract base class and dedicated VMExecution class, moved over execution logic from FullSynchronizer
