+++ packages/client/lib/sync/fullsync.ts
@@ -4,9 +4,7 @@ import { short } from '../util'
 import { Synchronizer, SynchronizerOptions } from './sync'
 import { BlockFetcher } from './fetcher'
 import { Block } from '@ethereumjs/block'
-import VM from '@ethereumjs/vm'
-import { DefaultStateManager } from '@ethereumjs/vm/dist/state'
-import { SecureTrie as Trie } from '@ethereumjs/trie'
+import { VMExecution } from './execution/vmexecution'
 
 /**
  * Implements an ethereum full sync synchronizer
@@ -14,108 +12,28 @@ import { SecureTrie as Trie } from '@ethereumjs/trie'
  */
 export class FullSynchronizer extends Synchronizer {
   private blockFetcher: BlockFetcher | null
-
-  public vm: VM
-  public runningBlocks: boolean
-
-  private stopSyncing: boolean
-  private vmPromise?: Promise<void | number>
+  private execution: VMExecution
 
   constructor(options: SynchronizerOptions) {
     super(options)
     this.blockFetcher = null
 
-    if (!this.config.vm) {
-      const trie = new Trie(this.stateDB)
-
-      const stateManager = new DefaultStateManager({
-        common: this.config.common,
-        trie,
-      })
-
-      this.vm = new VM({
-        common: this.config.common,
-        blockchain: this.chain.blockchain,
-        stateManager,
-      })
-    } else {
-      this.vm = this.config.vm
-      //@ts-ignore blockchain has readonly property
-      this.vm.blockchain = this.chain.blockchain
-    }
-
-    this.runningBlocks = false
-    this.stopSyncing = false
+    this.execution = new VMExecution({
+      config: options.config,
+      stateDB: options.stateDB,
+      chain: options.chain,
+    })
 
     const self = this
     this.chain.on('updated', async function () {
-      // for some reason, if we use .on('updated', this.runBlocks), it runs in the context of the Chain and not in the FullSync context..?
-      await self.runBlocks()
+      // for some reason, if we use .on('updated', this.runBlocks)
+      // it runs in the context of the Chain and not in the FullSync context..?
+      if (self.running) {
+        await self.execution.runBlocks()
+      }
     })
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
     this.chain.update()
-  }
-
-  /**
-   * This updates the VM once blocks were put in the VM
-   */
-  async runBlocks() {
-    if (!this.running || this.runningBlocks) {
-      return
-    }
-    this.runningBlocks = true
-    let blockCounter = 0
-    let txCounter = 0
-    const NUM_BLOCKS_PER_LOG_MSG = 50
-    try {
-      let oldHead = Buffer.alloc(0)
-      const newHeadBlock = await this.vm.blockchain.getHead()
-      let newHead = newHeadBlock.hash()
-      let firstHeadBlock = newHeadBlock
-      let lastHeadBlock = newHeadBlock
-      while (!newHead.equals(oldHead) && !this.stopSyncing) {
-        oldHead = newHead
-        this.vmPromise = this.vm.runBlockchain(this.vm.blockchain, 1)
-        const numExecuted = (await this.vmPromise) as number
-        if (numExecuted === 0) {
-          this.config.logger.warn(
-            `No blocks executed past chain head hash=${short(
-              newHead
-            )} number=${newHeadBlock.header.number.toNumber()}`
-          )
-          this.runningBlocks = false
-          return
-        }
-        const headBlock = await this.vm.blockchain.getHead()
-        newHead = headBlock.hash()
-        if (blockCounter === 0) {
-          firstHeadBlock = headBlock
-        }
-        // check if we did run a new block:
-        if (!newHead.equals(oldHead)) {
-          blockCounter += 1
-          txCounter += headBlock.transactions.length
-          lastHeadBlock = headBlock
-
-          if (blockCounter >= NUM_BLOCKS_PER_LOG_MSG) {
-            const firstNumber = firstHeadBlock.header.number.toNumber()
-            const firstHash = short(firstHeadBlock.hash())
-            const lastNumber = lastHeadBlock.header.number.toNumber()
-            const lastHash = short(lastHeadBlock.hash())
-            this.config.logger.info(
-              `Executed blocks count=${blockCounter} first=${firstNumber} hash=${firstHash} last=${lastNumber} hash=${lastHash} with txs=${txCounter}`
-            )
-            blockCounter = 0
-            txCounter = 0
-          }
-        }
-      }
-    } catch (error) {
-      this.emit('error', error)
-    } finally {
-      this.runningBlocks = false
-    }
-    return blockCounter
   }
 
   /**
@@ -254,16 +172,11 @@ export class FullSynchronizer extends Synchronizer {
    * @return {Promise}
    */
   async stop(): Promise<boolean> {
-    this.stopSyncing = true
-    if (this.vmPromise) {
-      // ensure that we wait that the VM finishes executing the block (and flushes the trie cache)
-      await this.vmPromise
-    }
-    await this.stateDB?.close()
-
     if (!this.running) {
       return false
     }
+    await this.execution.stop()
+
     if (this.blockFetcher) {
       this.blockFetcher.destroy()
       // TODO: Should this be deleted?
